Chore/#21 TabBar Font적용부 수정
